change shm type to long long

+++ include/timetrack.h
@@ -15,15 +15,15 @@
 typedef struct timeTrack{
     int eventID;
     int valid;
-    unsigned int array[TIME_COLUM];
+    long long array[TIME_COLUM];
 }timeTrack;
 
-unsigned int gettime_nanoTime(void)
+long long gettime_nanoTime(void)
 {
     struct timespec __tv;
     clock_gettime(CLOCK_MONOTONIC,&__tv);
     //clock_gettime(CLOCK_REALTIME,&__tv);
-    return(unsigned int)(__tv.tv_sec * 1e9 + __tv.tv_nsec);
+    return(long long)(__tv.tv_sec * 1e9 + __tv.tv_nsec);
 }
 
 struct fd_pair{
