+++ server/timetrack.h
@@ -12,7 +12,7 @@
 typedef struct timeTrack{
     int eventID;
     int valid;
-    unsigned int array[TIME_COLUM];
+    long long array[TIME_COLUM];
 }timeTrack;
 
 struct fd_pair{
